V1 done

+++ src/main/java/com/rs/utils/Configs.java
@@ -28,7 +28,7 @@ public class Configs {
         }
     }
 
-    public static long MAX_SIZE_EACH_TASK(){
+    public static int MAX_SIZE_EACH_TASK(){
         String size = prop.getProperty("max_size_each_task", "64k");
         int value = Integer.parseInt(size.substring(0, size.length()-1));
         if(size.endsWith("k") || size.endsWith("K")){
